Transfer AppRoutes to own file
