+++ src/context/UserContext.jsx
@@ -1,0 +1,3 @@
+import { createContext } from 'react';
+
+export const UserContext = createContext(null);
