+++ src/pages/Home/home.jsx
@@ -1,3 +1,5 @@
+import React from "react";
+
 export default function Home() {
   return (
     <div className="container--homepage">
